refactored data store handling
